+++ SuperBuild.cmake
@@ -196,6 +196,7 @@ option(USE_SYSTEM_SIRF "Build using an external version of SIRF" OFF)
 option(USE_SYSTEM_NIFTYREG "Build using an external version of NIFTYREG" OFF)
 option(USE_SYSTEM_GTest "Build using an external version of GTest" OFF)
 option(USE_SYSTEM_ACE "Build using an external version of ACE" ON)
+option(USE_SYSTEM_RocksDB "Build using an external version of RocksDB" ON)
 
 # SPM requires matlab
 if (BUILD_MATLAB)
@@ -299,6 +300,7 @@ if (BUILD_Gadgetron)
   list(APPEND ${PRIMARY_PROJECT_NAME}_DEPENDENCIES Gadgetron)
   set(Armadillo_REQUIRED_VERSION 4.600)
 endif()
+option(DISABLE_range-v3_TESTING "Disable range-v3 testing" ON)
 
 if (BUILD_siemens_to_ismrmrd)
   list(APPEND ${PRIMARY_PROJECT_NAME}_DEPENDENCIES siemens_to_ismrmrd)