+++ docker/build_gadgetron-ubuntu.sh
@@ -6,6 +6,7 @@ apt-get update -qq
 # boost is handled in build_system-ubuntu.sh
 apt-get install -yq --no-install-recommends \
   libhdf5-serial-dev     \
+  libboost-dev libboost-all-dev \
   libfftw3-dev           \
   h5utils                \
   jq                     \
@@ -19,7 +20,22 @@ apt-get install -yq --no-install-recommends \
   liblapack-dev          \
   liblapacke-dev         \
   libplplot-dev          \
-  libdcmtk-dev
+  libdcmtk-dev           \
+  libpugixml-dev         \
+  libgflags-dev          \
+  libssl-dev             \
+  libcurl4-openssl-dev   \
+  pkg-config             \
+  golang
+
+# install GCC9 required by Gadgetron
+apt-get install software-properties-common -y && \
+add-apt-repository ppa:ubuntu-toolchain-r/test -y && \
+apt-get update -y && \
+apt-get install gcc-9 g++-9 -y
+
+# apt-get install -y librocksdb-dev
+
 apt-get clean
 
 # install libgmock-dev if in the apt packages https://github.com/SyneRBI/SIRF-SuperBuild/issues/647#issuecomment-1042841986