+++ docker/ctest_sirf.sh
@@ -7,10 +7,10 @@ source ${INSTALL_DIR}/bin/env_sirf.sh
 export PATH=~/.local/bin:$PATH
 source ~/virtualenv/bin/activate 
 
-gadgetron >& gadgetron.log&
 # print for debugging
 cat $GITHUB_WORKSPACE/build/builds/SIRF/build/CMakeCache.txt
 cd $GITHUB_WORKSPACE/build
+gadgetron >& gadgetron.log&
 ctest --output-on-failure; test_fail=$?
 # echo "----------- Killing gadgetron server"
 # killall gadgetron