+++ version_config.cmake
@@ -29,11 +29,13 @@ if (APPLE) # really should be checking for CLang
     set(Boost_URL http://downloads.sourceforge.net/project/boost/boost/${Boost_VERSION}/boost_1_68_0.zip)
     set(Boost_MD5 f4096c4583947b0eb103c8539f1623a3)
 else()
+     # Use version 1.78.0 version
      set(Boost_VERSION 1.78.0)
      if (BUILD_GADGETRON)
+     # https://github.com/gadgetron/gadgetron/blob/12ffc43debb9bad2e170713006d29dea78d966bf/CMakeLists.txt#L205-L209
        set(Boost_REQUIRED_VERSION 1.71.0)
      else()
-       # Ubutnu 16.04 version should be fine
+       # Ubutnu 22.04 version should be fine
        set(Boost_REQUIRED_VERSION 1.58.0)
      endif()
      set(Boost_URL http://downloads.sourceforge.net/project/boost/boost/${Boost_VERSION}/boost_1_78_0.zip)
@@ -112,19 +114,17 @@ set(DEFAULT_NIFTYREG_TAG 8ad2f11507ddedb09ed74a9bd97377b70532ee75)
 set(NIFTYREG_REQUIRED_VERSION 1.5.68)
 
 ## ISMRMRD
+set(ISMRMRD_REQUIRED_VERSION "1.11.1")
 set(DEFAULT_ISMRMRD_URL https://github.com/ismrmrd/ismrmrd )
-set(DEFAULT_ISMRMRD_TAG v1.7.0)
-#if (WIN32)
-#  set(DEFAULT_ISMRMRD_URL https://github.com/SyneRBI/ismrmrd )
-#  set(DEFAULT_ISMRMRD_TAG program_options_fix)
-#else()
-#  set(DEFAULT_ISMRMRD_URL https://github.com/ismrmrd/ismrmrd )
-#  set(DEFAULT_ISMRMRD_TAG v1.4.1)
-#endif()
+set(DEFAULT_ISMRMRD_TAG v1.13.2)
+
+## siemens_to_ismrmrd
+set(DEFAULT_siemens_to_ismrmrd_URL https://github.com/ismrmrd/siemens_to_ismrmrd)
+set(DEFAULT_siemens_to_ismrmrd_TAG b87759e49e53dab4939147eb52b7a0e6465f3d04)
 
 ## Gadgetron
 set(DEFAULT_Gadgetron_URL https://github.com/gadgetron/gadgetron )
-set(DEFAULT_Gadgetron_TAG b6191eaaa72ccca6c6a5fe4c0fa3319694f512ab)
+set(DEFAULT_Gadgetron_TAG 42f11bf14b77b16f1ca5bcfbfa435d5ee8cb22a6)
 
 ## ASTRA
 set(DEFAULT_astra-toolbox_URL https://github.com/astra-toolbox/astra-toolbox )
@@ -154,11 +154,7 @@ set(DEFAULT_JSON_URL https://github.com/nlohmann/json.git )
 set(DEFAULT_JSON_TAG v3.10.4)
 
 # CCPi CIL
-# CIL-ASTRA has been merged into CIL from acf3ddf5c61b8e216fe7891d7720f9bbd436c9b3
-
-# minimum supported version of CIL supported is 22.0.0 or from commit acf3ddf5c61b8e216fe7891d7720f9bbd436c9b3
-# due to a change in CIL's building mechanism, however due to a unit test failure the minimum CIL required
-# commit is a6062410028c9872c5b355be40b96ed1497fed2a
+# minimum supported version of CIL supported is > 22.1.0 or from commit a6062410028c9872c5b355be40b96ed1497fed2a
 set(DEFAULT_CIL_URL https://github.com/TomographicImaging/CIL.git)
 set(DEFAULT_CIL_TAG a6062410028c9872c5b355be40b96ed1497fed2a)
 
@@ -172,6 +168,24 @@ set(DEFAULT_ROOT_TAG "v6-26-10")
 # ACE
 set(DEFAULT_ACE_URL https://github.com/paskino/libace-conda)
 set(DEFAULT_ACE_TAG v6.5.9)
+
+# range-v3
+set(DEFAULT_range-v3_URL https://github.com/ericniebler/range-v3.git )
+set(DEFAULT_range-v3_TAG 0.12.0)
+
+set(DEFAULT_RocksDB_URL https://github.com/facebook/rocksdb.git )
+set(DEFAULT_RocksDB_TAG v6.26.0)
+
+set(DEFAULT_mrd-storage-server_URL https://github.com/ismrmrd/mrd-storage-server.git)
+set(DEFAULT_mrd-storage-server_TAG origin/main)
+
+set(DEFAULT_Date_URL https://github.com/HowardHinnant/date.git )
+set(DEFAULT_Date_TAG master)
+
+# works only for Linux
+set(Go_URL https://go.dev/dl/go1.19.3.linux-amd64.tar.gz)
+set(Go_SHA256 74b9640724fd4e6bb0ed2a1bc44ae813a03f1e72a4c76253e2d5c015494430ba)
+
 
 option (DEVEL_BUILD "Developer Build" OFF)
 mark_as_advanced(DEVEL_BUILD)
@@ -192,11 +206,6 @@ if (DEVEL_BUILD)
   ## siemens_to_ismrmrd
   set(DEFAULT_siemens_to_ismrmrd_URL https://github.com/ismrmrd/siemens_to_ismrmrd )
   set(DEFAULT_siemens_to_ismrmrd_TAG b87759e49e53dab4939147eb52b7a0e6465f3d04)
-  if (BUILD_siemens_to_ismrmrd)
-    set(ISMRMRD_REQUIRED_VERSION "1.7")
-  else()
-    set(ISMRMRD_REQUIRED_VERSION "1.4.2.1")
-  endif()
 
   ## pet-rd-tools
   set(DEFAULT_pet_rd_tools_URL https://github.com/UCL/pet-rd-tools )
@@ -205,6 +214,13 @@ if (DEVEL_BUILD)
   # CCPi CIL
   set(DEFAULT_CIL_URL https://github.com/TomographicImaging/CIL.git)
   set(DEFAULT_CIL_TAG origin/master )
+
+  # Gadgetron
+  set(DEFAULT_Gadgetron_TAG origin/master)
+  
+  # ismrmrd
+  set(DEFAULT_ISMRMRD_TAG origin/master)
+
   
 
 else()
@@ -213,12 +229,6 @@ else()
   ## STIR
   set(DEFAULT_STIR_URL https://github.com/UCL/STIR )
   set(DEFAULT_STIR_TAG rel_5.1.0)
-
-  # ismrmrd
-  set(DEFAULT_ISMRMRD_TAG v1.7.0)
-  ## siemens_to_ismrmrd
-  set(DEFAULT_siemens_to_ismrmrd_URL https://github.com/ismrmrd/siemens_to_ismrmrd)
-  set(DEFAULT_siemens_to_ismrmrd_TAG b87759e49e53dab4939147eb52b7a0e6465f3d04)
 
   ## pet-rd-tools
   set(DEFAULT_pet_rd_tools_URL https://github.com/UCL/pet-rd-tools )
@@ -296,6 +306,18 @@ set(SPM_TAG ${DEFAULT_SPM_TAG} CACHE STRING ON)
 set(JSON_URL ${DEFAULT_JSON_URL} CACHE STRING ON)
 set(JSON_TAG ${DEFAULT_JSON_TAG} CACHE STRING ON)
 
+set(range-v3_URL ${DEFAULT_range-v3_URL} CACHE STRING ON)
+set(range-v3_TAG ${DEFAULT_range-v3_TAG} CACHE STRING ON)
+
+set(RocksDB_URL ${DEFAULT_RocksDB_URL} CACHE STRING ON)
+set(RocksDB_TAG ${DEFAULT_RocksDB_TAG} CACHE STRING ON)
+
+set(mrd-storage-server_URL ${DEFAULT_mrd-storage-server_URL} CACHE STRING ON)
+set(mrd-storage-server_TAG ${DEFAULT_mrd-storage-server_TAG} CACHE STRING ON)
+
+set(Date_URL ${DEFAULT_Date_URL} CACHE STRING ON)
+set(Date_TAG ${DEFAULT_Date_TAG} CACHE STRING ON)
+
 mark_as_advanced(SIRF_URL SIRF_TAG STIR_URL STIR_TAG
   Gadgetron_URL Gadgetron_TAG
   siemens_to_ismrmrd_URL siemens_to_ismrmrd_TAG
@@ -311,9 +333,12 @@ mark_as_advanced(SIRF_URL SIRF_TAG STIR_URL STIR_TAG
   ITK_URL ITK_TAG
   SPM_URL SPM_TAG
   JSON_URL JSON_TAG
+  range-v3_URL range-v3_TAG
   ROOT_URL ROOT_TAG
   astra-toolbox_URL astra-toolbox_TAG
   astra-python-wrapper_URL astra-python-wrapper_TAG
   ACE_URL ACE_TAG
-  
+  RocksDB_URL RocksDB_TAG
+  mrd-storage-server_URL mrd-storage-server_TAG
+  Date_URL Date_TAG
 )